Change type of function argument.

+++ cpp/subprojects/boosting/src/boosting/statistics/statistics_example_wise_common.hpp
@@ -283,7 +283,7 @@ namespace boosting {
             /**
              * @see `IHistogram::addToBin`
              */
-            void addToBin(uint32 binIndex, uint32 statisticIndex, uint32 weight) override {
+            void addToBin(uint32 binIndex, uint32 statisticIndex, float64 weight) override {
                 this->statisticViewPtr_->addToRow(binIndex, originalStatisticView_.gradients_row_cbegin(statisticIndex),
                                                   originalStatisticView_.gradients_row_cend(statisticIndex),
                                                   originalStatisticView_.hessians_row_cbegin(statisticIndex),
